finished

+++ routes/web.php
@@ -20,21 +20,21 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+Route::get('/dashboard', [StudentController::class, 'index'])
+    ->middleware(['auth', 'handle.roles:admin']) 
+    ->name('dashboard');
 
-Route::get('/student', function () {
-    return view('student');
-})->middleware(['auth', 'verified'])->name('student');
+Route::get('/userdashboard', [StudentController::class, 'userDashboard'])
+    ->middleware(['auth', 'verified'])
+    ->name('userdashboard');
 
-Route::get('/students', [StudentController::class, 'index'])->name('students.index');
-Route::get('/students/create', [StudentController::class, 'create'])->name('students.create');
-Route::post('/students', [StudentController::class, 'store'])->name('students.store');
-Route::get('/students/{id}', [StudentController::class, 'show'])->name('students.show');
-Route::get('/students/{id}/edit', [StudentController::class, 'edit'])->name('students.edit');
-Route::put('/students/{id}', [StudentController::class, 'update'])->name('students.update');
-Route::delete('/students/{id}', [StudentController::class, 'destroy'])->name('students.destroy');
+// Route::get('/students', [StudentController::class, 'index'])->middleware(['auth', 'role:admin'])->name('students.index');
+Route::get('/students/create', [StudentController::class, 'create'])->middleware(['auth', 'role:admin'])->name('students.create');
+Route::post('/students', [StudentController::class, 'store'])->middleware(['auth', 'role:admin'])->name('students.store');
+Route::get('/students/{id}', [StudentController::class, 'show'])->middleware(['auth', 'role:admin'])->name('students.show');
+Route::get('/students/{id}/edit', [StudentController::class, 'edit'])->middleware(['auth', 'role:admin'])->name('students.edit');
+Route::put('/students/{id}', [StudentController::class, 'update'])->middleware(['auth', 'role:admin'])->name('students.update');
+Route::delete('/students/{id}', [StudentController::class, 'destroy'])->middleware(['auth', 'role:admin'])->name('students.destroy');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
